added nb_nodes nb_edges

+++ LSBP/data/cache.py
@@ -12,6 +12,7 @@ import numpy as np
 
 from LSBP.data.load import get_project_root, json2dict, dict2json
 from LSBP.utils import Bunch
+from LSBP.utils.check import check_empty_dir
 
 class Cache:
     ''' Utility for cache. '''
@@ -20,15 +21,15 @@ class Cache:
         
         data_dir = os.path.basename(directory).split('.')[0]
         self.directory = os.path.join(get_project_root(), 'preproc_data', data_dir)
-        self.is_empty = not(os.path.exists(self.directory))
+        self.is_empty = check_empty_dir(self.directory)
 
         if not use_cache:
             self.clear()
 
         if not self.is_empty:
-            print(f'Use cached data in {self.directory}')
+            print(f'Use cached data in {self.directory}.')
         else:
-            os.makedirs(self.directory)
+            os.makedirs(self.directory, exist_ok=True)
 
     def add(self, graph: Bunch):
         ''' Cache graph information on disk.
@@ -41,9 +42,12 @@ class Cache:
         # Save information into cache
         dict2json(graph.in_degrees, os.path.join(self.directory, 'in_degrees.json'))
         dict2json(graph.out_degrees, os.path.join(self.directory, 'out_degrees.json'))
-        np.save(os.path.join(self.directory, 'nodes.npy'), graph.nodes)
+        #np.save(os.path.join(self.directory, 'nodes.npy'), graph.nodes)
         dict2json(graph.label2idx, os.path.join(self.directory, 'label2idx.json'))
         dict2json(graph.idx2label, os.path.join(self.directory, 'idx2label.json'))
+
+        stats = {'nb_nodes': graph.nb_nodes, 'nb_edges': graph.nb_edges}
+        dict2json(stats, os.path.join(self.directory, 'stats.json'))
 
     def load(self) -> Bunch:
         ''' Load cached information into a Bunch object. 
@@ -54,14 +58,16 @@ class Cache:
 
         graph = Bunch()
 
+        print('Loading data from cache...')
         # Load information from cache
         graph.outdir = self.directory
         graph.label2idx = json2dict(os.path.join(self.directory, 'label2idx.json'))
-        graph.idx2label = json2dict(os.path.join(self.directory, 'idx2label.json'))
+        graph.idx2label = json2dict(os.path.join(self.directory, 'idx2label.json'), keys_int=True)
         graph.in_degrees = json2dict(os.path.join(self.directory, 'in_degrees.json'), keys_int=True)
         graph.out_degrees = json2dict(os.path.join(self.directory, 'out_degrees.json'), keys_int=True)
-        graph.nodes = np.load(os.path.join(self.directory, 'nodes.npy'))
-        graph.nb_nodes = len(graph.nodes)
+        #graph.nodes = np.load(os.path.join(self.directory, 'nodes.npy'))
+        stats = json2dict(os.path.join(self.directory, 'stats.json'))
+        graph.nb_nodes, graph.nb_edges = stats.get('nb_nodes'), stats.get('nb_edges')
         graph.files = [f for f in os.listdir(self.directory) if os.path.basename(self.directory) in f]
 
         return graph
@@ -69,5 +75,6 @@ class Cache:
     def clear(self):
         ''' Clear cache by deleting directory. '''
 
-        shutil.rmtree(self.directory)
+        if os.path.exists(self.directory):
+            shutil.rmtree(self.directory)
         self.is_empty = True
